Change PowerShell installation for Ubuntu 24.04

+++ scripts/ubuntu/programs/powershell.sh
@@ -4,25 +4,39 @@
 
 echo "Installing PowerShell..."
 
-# Update the list of packages
-sudo apt-get update
+# Get the Ubuntu version
+version=$(lsb_release -rs)
 
-# Install pre-requisite packages.
-sudo apt-get install -y wget apt-transport-https software-properties-common
+# Check if the version is greater than or equal to 24.04
+if [ $(echo "$version 24.04" | awk '{print ($1 >= $2)}') -ne 0 ]; then
+    # Installation method for Ubuntu 24.04 or greater
 
-# Download the Microsoft repository GPG keys
-OS_RELEASE=$(lsb_release -rs)
-MS_REPOSITORY_URL="https://packages.microsoft.com/config/ubuntu/$OS_RELEASE/packages-microsoft-prod.deb"
-wget -q $MS_REPOSITORY_URL -O ~/packages-microsoft-prod.deb
+    # https://github.com/PowerShell/PowerShell/issues/21385
+    # Using snap as an alternative
+    sudo snap install powershell --classic
+else
+    # Installation method for the script for versions less than Ubuntu 24.04
 
-# Register the Microsoft repository GPG keys
-sudo dpkg -i ~/packages-microsoft-prod.deb
+    # Update the list of packages
+    sudo apt-get update
 
-# Cleanup the deb package
-rm ~/packages-microsoft-prod.deb
+    # Install pre-requisite packages.
+    sudo apt-get install -y wget apt-transport-https software-properties-common
 
-# Update the list of products
-sudo apt-get update
+    # Download the Microsoft repository GPG keys
+    OS_RELEASE=$(lsb_release -rs)
+    MS_REPOSITORY_URL="https://packages.microsoft.com/config/ubuntu/$OS_RELEASE/packages-microsoft-prod.deb"
+    wget -q $MS_REPOSITORY_URL -O ~/packages-microsoft-prod.deb
 
-# Install PowerShell
-sudo apt-get install -y powershell
+    # Register the Microsoft repository GPG keys
+    sudo dpkg -i ~/packages-microsoft-prod.deb
+
+    # Cleanup the deb package
+    rm ~/packages-microsoft-prod.deb
+
+    # Update the list of products
+    sudo apt-get update
+
+    # Install PowerShell
+    sudo apt-get install -y powershell
+fi
